real_scraper/requirements.txt
python-dotenv
Flask
langchain
openai
wikipedia
pyowm
validators
requests
json
beautifulsoup4
tiktoken